// proto/cosmoshub/ts/v25.2.0/cosmos/autocli/v1/options.ts
export interface ModuleOptions {
  /** tx describes the tx commands for the module. */
  tx: ServiceCommandDescriptor | undefined;
  /** query describes the queries commands for the module. */
  query: ServiceCommandDescriptor | undefined;
}

/** ServiceCommandDescriptor describes a CLI command based on a protobuf service. */
export interface ServiceCommandDescriptor {
  /**
   * service is the fully qualified name of the protobuf service to build
   * the command from. It can be left empty if sub_commands are used instead
   * which may be the case if a module provides multiple tx and/or query services.
   */
  service: string;
  /**
   * rpc_command_options are options for commands generated from rpc methods.
   * If no options are specified for a given rpc method on the service, a
   * command will be generated for that method with the default options.
   */
  rpcCommandOptions: RpcCommandOptions[];
  /**
   * sub_commands is a map of optional sub-commands for this command based on
   * different protobuf services. The map key is used as the name of the
   * sub-command.
   */
  subCommands: { [key: string]: ServiceCommandDescriptor };
  /**
   * enhance_custom_commands specifies whether to skip the service when generating commands, if a custom command already
   * exists, or enhance the existing command. If set to true, the custom command will be enhanced with the services from
   * gRPC. otherwise when a custom command exists, no commands will be generated for the service.
   */
  enhanceCustomCommand: boolean;
  /** short is an optional parameter used to override the short description of the auto generated command. */
  short: string;
}

export interface ServiceCommandDescriptor_SubCommandsEntry {
  key: string;
  value: ServiceCommandDescriptor | undefined;
}

/**
 * RpcCommandOptions specifies options for commands generated from protobuf
 * rpc methods.
 */
export interface RpcCommandOptions {
  /**
   * rpc_method is short name of the protobuf rpc method that this command is
   * generated from.
   */
  rpcMethod: string;
  /**
   * use is the one-line usage method. It also allows specifying an alternate
   * name for the command as the first word of the usage text.
   *
   * By default the name of an rpc command is the kebab-case short name of the
   * rpc method.
   */
  use: string;
  /** long is the long message shown in the 'help <this-command>' output. */
  long: string;
  /** short is the short description shown in the 'help' output. */
  short: string;
  /** example is examples of how to use the command. */
  example: string;
  /** alias is an array of aliases that can be used instead of the first word in Use. */
  alias: string[];
  /**
   * suggest_for is an array of command names for which this command will be suggested -
   * similar to aliases but only suggests.
   */
  suggestFor: string[];
  /** deprecated defines, if this command is deprecated and should print this string when used. */
  deprecated: string;
  /**
   * version defines the version for this command. If this value is non-empty and the command does not
   * define a "version" flag, a "version" boolean flag will be added to the command and, if specified,
   * will print content of the "Version" variable. A shorthand "v" flag will also be added if the
   * command does not define one.
   */
  version: string;
  /**
   * flag_options are options for flags generated from rpc request fields.
   * By default all request fields are configured as flags. They can
   * also be configured as positional args instead using positional_args.
   */
  flagOptions: { [key: string]: FlagOptions };
  /** positional_args specifies positional arguments for the command. */
  positionalArgs: PositionalArgDescriptor[];
  /** skip specifies whether to skip this rpc method when generating commands. */
  skip: boolean;
  /**
   * gov_proposal specifies whether autocli should generate a gov proposal transaction for this rpc method.
   * Normally autocli generates a transaction containing the message and broadcast it.
   * However, when true, autocli generates a proposal transaction containing the message and broadcast it.
   * This option is ineffective for query commands.
   */
  govProposal: boolean;
}
// ...
const baseModuleOptions: object = {};

export const ModuleOptions = {
  encode(
    message: ModuleOptions,
    writer: _m0.Writer = _m0.Writer.create()
  ): _m0.Writer {
    if (message.tx !== undefined) {
      ServiceCommandDescriptor.encode(
        message.tx,
        writer.uint32(10).fork()
      ).ldelim();
    }
    if (message.query !== undefined) {
      ServiceCommandDescriptor.encode(
        message.query,
        writer.uint32(18).fork()
      ).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): ModuleOptions {
    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = { ...baseModuleOptions } as ModuleOptions;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.tx = ServiceCommandDescriptor.decode(reader, reader.uint32());
          break;
        case 2:
          message.query = ServiceCommandDescriptor.decode(
            reader,
            reader.uint32()
          );
          break;
        default:
          reader.skipType(tag & 7);
          break;
      }
    }
    return message;
  },

  fromJSON(object: any): ModuleOptions {
    const message = { ...baseModuleOptions } as ModuleOptions;
    if (object.tx !== undefined && object.tx !== null) {
      message.tx = ServiceCommandDescriptor.fromJSON(object.tx);
    } else {
      message.tx = undefined;
    }
    if (object.query !== undefined && object.query !== null) {
      message.query = ServiceCommandDescriptor.fromJSON(object.query);
    } else {
      message.query = undefined;
    }
    return message;
  },

  toJSON(message: ModuleOptions): unknown {
    const obj: any = {};
    message.tx !== undefined &&
      (obj.tx = message.tx
        ? ServiceCommandDescriptor.toJSON(message.tx)
        : undefined);
    message.query !== undefined &&
      (obj.query = message.query
        ? ServiceCommandDescriptor.toJSON(message.query)
        : undefined);
    return obj;
  },

  fromPartial(object: DeepPartial<ModuleOptions>): ModuleOptions {
    const message = { ...baseModuleOptions } as ModuleOptions;
    if (object.tx !== undefined && object.tx !== null) {
      message.tx = ServiceCommandDescriptor.fromPartial(object.tx);
    } else {
      message.tx = undefined;
    }
    if (object.query !== undefined && object.query !== null) {
      message.query = ServiceCommandDescriptor.fromPartial(object.query);
    } else {
      message.query = undefined;
    }
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | undefined;
export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends Array<infer U>
  ? Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U>
  ? ReadonlyArray<DeepPartial<U>>
  : T extends {}
  ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;